feat: Implement WithTimeout for both embassy and tokio
Add implementations of the WithTimeout trait for both embassy and tokio runtimes. This allows asynchronous operations to be executed with a specified timeout.
Ensure that both features cannot be enabled simultaneously to prevent conflicts.
Update Cargo.toml to include optional dependencies for embassy-time and tokio.
